requirements.txt
Pillow>=10.2.0
opencv-python>=4.8.1.78
mediapipe>=0.10.0
numpy>=1.24.0
